+++ Assets/TechC/VBattle/Scripts/InGame/Player/AttackBase.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9370bbb09989c8549874b9f16b718f63
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
